requirements.txt
absl-py==0.13.0
astunparse==1.6.3
blis==0.7.4
Brotli==1.0.9
cachetools==4.2.2
catalogue==2.0.4
certifi==2021.5.30
charset-normalizer==2.0.3
click==7.1.2
connectorx==0.2.0
cycler==0.10.0
cymem==2.0.5
dash==1.21.0
dash-core-components==1.17.1
dash-html-components==1.1.4
dash-table==4.12.0
en-core-web-lg==3.1.0
et-xmlfile==1.1.0
Flask==2.0.1
Flask-Compress==1.10.1
flatbuffers==1.12
future==0.18.2
gast==0.4.0
geographiclib==1.52
geopy==2.2.0
google-auth==1.33.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
murmurhash==1.0.5
nltk==3.6.2
numpy==1.19.5
oauthlib==3.1.1
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
pathy==0.6.0
Pillow==8.3.1
plotly==5.1.0
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.7.0
six==1.15.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
srsly==2.4.1
tenacity==8.0.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
thinc==8.0.8
tqdm==4.61.2
tweepy==3.10.0
typer==0.3.2
typing-extensions==3.7.4.3
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
wordcloud==1.8.1
wrapt==1.12.1
